+++ docs/img_convert.sh
@@ -1,8 +1,9 @@
+#!/bin/bash
 
-convert -density 120 tensors_logo.svg tensors_logo_120.png 
-convert -density 144 tensors_logo.svg tensors_logo_144.png 
+convert -background 'rgba(255,255,255,0)' -density 120 mpnum_logo.svg mpnum_logo_120.png 
+convert -background 'rgba(255,255,255,0)' -density 144 mpnum_logo.svg mpnum_logo_144.png 
 
 for i in tensors_*; do
-  convert -density 100 $i ${i/.svg/.png}
+  convert -background 'rgba(255,255,255,0)' -density 135 "$i" "${i/.svg/.png}"
 done