+++ mpnum/utils/__init__.py
@@ -1,4 +1,4 @@
 # encoding: utf-8
 
-from .array_transforms import *
-from .extmath import *
+from .array_transforms import *  # noqa: F401, F403
+from .extmath import *           # noqa: F401, F403